% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnutils.R
\name{chg_col_names}
\alias{chg_col_names}
\title{Standardize col names of a data frame}
\usage{
chg_col_names(df)
}
\arguments{
\item{df}{data frame to prettify col names}
}
\value{
same data frame with new col names
}
\description{
Create clean and ease to column names. Replaces punctuation with underscore and brings to lower.
}
\examples{
chg_col_names(data.frame("first.col" = c(1, 2), "second col" = c(3, 4)))

}
