% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_structure.R
\name{create_package_url}
\alias{create_package_url}
\title{Create URL from summary}
\usage{
create_package_url(...)
}
\arguments{
\item{...}{filtering arguments. otherwise will mutate the full table}
}
\value{
dataframe
}
\description{
Using the summary file create URL to download packages.
It is designed to be used in a pipeline with retrieve_info.
}
\examples{
create_package_url()

}
