% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_to_df.R
\name{download_files}
\alias{download_files}
\title{Download one or multiple URLs}
\usage{
download_files(urls, exp_jsn = FALSE, json_col = NULL, keep_nodes = NULL)
}
\arguments{
\item{urls}{list, vector or character. the addresses where the tables or zip file exists}

\item{exp_jsn}{logical. expand json string if found in the data frame}

\item{json_col}{if data frame is provided, specify column where json resides}

\item{keep_nodes}{subset the nodes from json. vector or single value}
}
\value{
list of data frames
}
\description{
Loops over a series of urls, download the content zip or csv and transform it into a data frame.
}
\examples{
url1 <- "https://ipchem.jrc.ec.europa.eu/public/EMPODAT/2015/IPCHEM_EMPODAT_a008a910-dda6-4e0d-a298-014c6eb1973c.zip"
url2 <- "https://ipchem.jrc.ec.europa.eu/public/EMPODAT/2016/IPCHEM_EMPODAT_c0ee2c5e-16c6-45c5-a572-dbf04d5abf26.zip"

ls_url <- c(url1, url2)

download_files(ls_url)

}
