% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnutils.R
\name{expand_args}
\alias{expand_args}
\title{Parse ellipsis}
\usage{
expand_args(...)
}
\arguments{
\item{...}{vector, list, data frame}
}
\value{
a list of arguments
}
\description{
Transform arguments into a list.
Unnamed arguments are eliminated.
If a data frame is argument, each column will become a list.
}
\examples{
expand_args(x = c("a", "b"), c(2), data.frame("col1" = c(1, 2)))

}
