% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_util.R
\name{expand_json}
\alias{expand_json}
\title{Parse json into data frame}
\usage{
expand_json(object, json_col = NULL, keep_nodes = NULL)
}
\arguments{
\item{object}{a data frame or a vector that contains json files}

\item{json_col}{if data frame is provided, specify column where json resides}

\item{keep_nodes}{subset the nodes from json. vector or single value}
}
\value{
data.table
}
\description{
Explode a json text stored in a vector or a columns from data frame.
If part of a data frame, the nodes will be transformed in columns.
}
\examples{
df <- data.frame("x" = c(1), "y" = c("{\"a\": 1, \"b\": 2}"))
expand_json(df, json_col = "y")
}
