% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_csv_content}
\alias{get_csv_content}
\title{Download csv file from url}
\usage{
get_csv_content(url)
}
\arguments{
\item{url}{the url must point to a csv file. return error otherwise}
}
\value{
dataframe object
}
\description{
Download a CSV file from a url and tranform it into a data frame.
}
\examples{
get_csv_content("https://ipchem.jrc.ec.europa.eu/public/IPCHEM_public_dataset_packages.csv")

}
