% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_json_content}
\alias{get_json_content}
\title{Download json file from url}
\usage{
get_json_content(url)
}
\arguments{
\item{url}{the url must point to a json file. return error otherwise}
}
\value{
list object
}
\description{
Download a json file from a url and transform it into a list.
}
\examples{
get_json_content("https://ipchem.jrc.ec.europa.eu/public/AIRBASE/IPCheM_AIRBASE_metadata.json")

}
