% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_raw_content}
\alias{get_raw_content}
\title{Download raw data from a URL}
\usage{
get_raw_content(url)
}
\arguments{
\item{url}{address of the site}
}
\value{
The binary body response
}
\description{
Designed to work with content in a binary form.
}
\examples{
get_raw_content("https://ipchem.jrc.ec.europa.eu/public/ESBUBA/2005/IPCHEM_ESBUBA_c3f8c698-52f2-408b-8138-35e5913a6029.zip")

}
