% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnutils.R
\name{get_url_suffix}
\alias{get_url_suffix}
\title{Show url last part}
\usage{
get_url_suffix(url)
}
\arguments{
\item{url}{character}
}
\value{
character; sufix
}
\description{
Extract the last part of an url, that being and extension or just folder path.
}
\examples{
get_url_suffix("abc.com/file.csv")

}
