% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnutils.R
\name{isin}
\alias{isin}
\title{Check if a pattern is in a string}
\usage{
isin(var, pattern, type = "any")
}
\arguments{
\item{var}{string, vector of strings, list of strings}

\item{pattern}{string, vector of strings, list of strings. Not regular expressions.}

\item{type}{string; "any" or "all". Check if any or all values from pattern are part of var.}
}
\value{
bool; vector of equal length as var
}
\description{
Used in filter argument for retrieve_info or equivalent function
}
\examples{
isin("abc", "a")

}
