% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_util.R
\name{json_to_df}
\alias{json_to_df}
\title{Show json structure}
\usage{
json_to_df(jsn, pprint = FALSE)
}
\arguments{
\item{jsn}{a string in json format}

\item{pprint}{show a nice print of the json}
}
\value{
data frame with information
}
\description{
Expand and show the json structure in an easy to use and see format.
}
\examples{
jsn <- "{\"a\": 1, \"b\": \"horse\"}"
json_to_df(jsn, pprint = TRUE)

}
