% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_structure.R
\name{load_summary}
\alias{load_summary}
\title{Read the summary file from IPCHEM site}
\usage{
load_summary()
}
\value{
a data frame called chem_summary stored in the environment
}
\description{
The file is stored in the local environment and used by retrieve_info function.
Do not alter the downloaded file.
Otherwise, rerun this function.
}
\examples{
load_summary()

}
