% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnutils.R
\name{pprint_vect}
\alias{pprint_vect}
\title{Prettify a vector of values}
\usage{
pprint_vect(v)
}
\arguments{
\item{v}{vector of values}
}
\description{
Prints the elements in alphabetical order with index letter.
}
\examples{
pprint_vect(c("Gabriel", "Inzaghi"))

}
