% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zip.R
\name{read_zip_pack}
\alias{read_zip_pack}
\title{Unzip and extract CSV}
\usage{
read_zip_pack(raw_response)
}
\arguments{
\item{raw_response}{the response of the get_content function}
}
\value{
data frame
}
\description{
Unzip a csv file from a raw response and transform into a data frame.
The archive must contain a csv, otherwise will return NULL.
}
