% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_structure.R
\name{retrieve_cas_number}
\alias{retrieve_cas_number}
\title{Wrapper around retrieve_info}
\usage{
retrieve_cas_number(..., filter)
}
\arguments{
\item{...}{vector or data frame used for filtering the results}

\item{filter}{filter the target using regular expression}
}
\value{
data frame with the first column as the target and all the other columns from filters
}
\description{
Specific case of retrieve_info for target = cas_number
}
\examples{
retrieve_cas_number(filter = (cas_number \%in\% c("83003-12-7", "144171-61-9")))

}
