% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_structure.R
\name{retrieve_dataset}
\alias{retrieve_dataset}
\title{Wrapper around retrieve_info}
\usage{
retrieve_dataset(..., filter)
}
\arguments{
\item{...}{vector or data frame used for filtering the results}

\item{filter}{filter the target using regular expression}
}
\value{
data frame with the first column as the target and all the other columns from filters
}
\description{
Specific case of retrieve_info for target = dataset
}
\examples{
retrieve_dataset(filter = stringr::str_detect(dataset, "^EMP"))

}
