% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_structure.R
\name{retrieve_info}
\alias{retrieve_info}
\title{Explore the summary file}
\usage{
retrieve_info(..., target, filter)
}
\arguments{
\item{...}{vector or data frame used for filtering the results}

\item{target}{using the filters, the column that will be displayed for available values}

\item{filter}{filter the target using regular expression}
}
\value{
data frame with the first column as the target and all the other columns from filters
}
\description{
Works with the file generated by load_summary().
The function starts from one variable and builds upon it. It is the reverse way of filtering rows.
}
\examples{
retrieve_info(target = media, filter = stringr::str_detect(media, "Water"))

}
