% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_structure.R
\name{retrieve_meta}
\alias{retrieve_meta}
\title{Get metadata info}
\usage{
retrieve_meta(dataset)
}
\arguments{
\item{dataset}{character or vector with names of datasets}
}
\value{
list
}
\description{
Using the name of a dataset bring the json metadata file.
}
\examples{
retrieve_meta("NAIADES")

}
