% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_to_df.R
\name{url_to_df}
\alias{url_to_df}
\title{Utility to download a data frame from url}
\usage{
url_to_df(url, exp_jsn = FALSE, json_col = NULL, keep_nodes = NULL)
}
\arguments{
\item{url}{character; the address where the table or zip file exists}

\item{exp_jsn}{logical. expand json string if found in the data frame}

\item{json_col}{if data frame is provided, specify column where json resides}

\item{keep_nodes}{subset the nodes from json. vector or single value}
}
\value{
data frame
}
\description{
Using a url, download the content zip or csv and transform it into a data frame.
}
\examples{
url_to_df("https://ipchem.jrc.ec.europa.eu/public/ESBUBA/2005/IPCHEM_ESBUBA_c3f8c698-52f2-408b-8138-35e5913a6029.zip")

}
